Extract X509Stack and X509StoreCtx methods for better sharing
