+++ src/Common/src/Interop/Unix/System.Security.Cryptography.Native/Interop.Crypto.cs
@@ -66,34 +66,8 @@ internal static partial class Interop
         private static extern int GetAsn1StringBytes(IntPtr asn1, byte[] buf, int cBuf);
 
         [DllImport(Libraries.CryptoNative)]
-        internal static extern SafeX509StackHandle NewX509Stack();
-
-        [DllImport(Libraries.CryptoNative)]
-        internal static extern int GetX509StackFieldCount(SafeX509StackHandle stack);
-
-        [DllImport(Libraries.CryptoNative)]
-        internal static extern int GetX509StackFieldCount(SafeSharedX509StackHandle stack);
-
-        /// <summary>
-        /// Gets a pointer to a certificate within a STACK_OF(X509). This pointer will later
-        /// be freed, so it should be cloned via new X509Certificate2(IntPtr)
-        /// </summary>
-        [DllImport(Libraries.CryptoNative)]
-        internal static extern IntPtr GetX509StackField(SafeX509StackHandle stack, int loc);
-
-        /// <summary>
-        /// Gets a pointer to a certificate within a STACK_OF(X509). This pointer will later
-        /// be freed, so it should be cloned via new X509Certificate2(IntPtr)
-        /// </summary>
-        [DllImport(Libraries.CryptoNative)]
-        internal static extern IntPtr GetX509StackField(SafeSharedX509StackHandle stack, int loc);
-
-        [DllImport(Libraries.CryptoNative)]
         [return: MarshalAs(UnmanagedType.Bool)]
         internal static extern bool PushX509StackField(SafeX509StackHandle stack, SafeX509Handle x509);
-
-        [DllImport(Libraries.CryptoNative)]
-        internal static extern void RecursiveFreeX509Stack(IntPtr stack);
 
         internal static string GetX509RootStorePath()
         {
