+++ src/Common/src/Microsoft/Win32/SafeHandles/SafeX509Handles.Unix.cs
@@ -73,66 +73,4 @@ namespace Microsoft.Win32.SafeHandles
             get { return handle == IntPtr.Zero; }
         }
     }
-
-    [SecurityCritical]
-    internal sealed class SafeX509StoreCtxHandle : SafeHandle
-    {
-        private SafeX509StoreCtxHandle() :
-            base(IntPtr.Zero, ownsHandle: true)
-        {
-        }
-
-        internal SafeX509StoreCtxHandle(IntPtr handle, bool ownsHandle) :
-            base(handle, ownsHandle)
-        {
-        }
-
-        protected override bool ReleaseHandle()
-        {
-            Interop.Crypto.X509StoreCtxDestroy(handle);
-            SetHandle(IntPtr.Zero);
-            return true;
-        }
-
-        public override bool IsInvalid
-        {
-            get { return handle == IntPtr.Zero; }
-        }
-    }
-
-    [SecurityCritical]
-    internal sealed class SafeX509StackHandle : SafeHandle
-    {
-        private SafeX509StackHandle() :
-            base(IntPtr.Zero, ownsHandle: true)
-        {
-        }
-
-        protected override bool ReleaseHandle()
-        {
-            Interop.Crypto.RecursiveFreeX509Stack(handle);
-            SetHandle(IntPtr.Zero);
-            return true;
-        }
-
-        public override bool IsInvalid
-        {
-            get { return handle == IntPtr.Zero; }
-        }
-    }
-
-    /// <summary>
-    /// Represents access to a STACK_OF(X509)* which is a member of a structure tracked
-    /// by another SafeHandle.
-    /// </summary>
-    [SecurityCritical]
-    internal sealed class SafeSharedX509StackHandle : SafeInteriorHandle
-    {
-        internal static readonly SafeSharedX509StackHandle InvalidHandle = new SafeSharedX509StackHandle();
-
-        private SafeSharedX509StackHandle() :
-            base(IntPtr.Zero, ownsHandle: true)
-        {
-        }
-    }
 }
